refactor: reorganize animation utilities and update imports
- Moved animation utilities to a new directory for better structure.
- Updated import paths in components to reflect the new location of animation utilities.
- Removed the old animation-utils file to streamline the codebase.
