--- src/lib/animation-utils.ts
@@ -1,175 +1,0 @@
-'use client'
-
-import type { Variants } from 'motion/react'
-
-// Staggered children animation
-export const staggerContainer: Variants = {
-  hidden: { opacity: 0 },
-  show: {
-    opacity: 1,
-    transition: {
-      staggerChildren: 0.1,
-    },
-  },
-}
-
-// Fade up animation for individual elements
-export const fadeUp: Variants = {
-  hidden: {
-    opacity: 0,
-    y: 20,
-  },
-  show: {
-    opacity: 1,
-    y: 0,
-    transition: {
-      type: 'spring',
-      damping: 25,
-      stiffness: 200,
-    },
-  },
-}
-
-// Fade in animation
-export const fadeIn: Variants = {
-  hidden: { opacity: 0 },
-  show: {
-    opacity: 1,
-    transition: {
-      duration: 0.5,
-    },
-  },
-}
-
-// Scale animation
-export const scaleUp: Variants = {
-  hidden: {
-    opacity: 0,
-    scale: 0.95,
-  },
-  show: {
-    opacity: 1,
-    scale: 1,
-    transition: {
-      type: 'spring',
-      damping: 25,
-      stiffness: 200,
-    },
-  },
-}
-
-// Slide in from left
-export const slideInLeft: Variants = {
-  hidden: {
-    opacity: 0,
-    x: -30,
-  },
-  show: {
-    opacity: 1,
-    x: 0,
-    transition: {
-      type: 'spring',
-      damping: 25,
-      stiffness: 200,
-    },
-  },
-}
-
-// Slide in from right
-export const slideInRight: Variants = {
-  hidden: {
-    opacity: 0,
-    x: 30,
-  },
-  show: {
-    opacity: 1,
-    x: 0,
-    transition: {
-      type: 'spring',
-      damping: 25,
-      stiffness: 200,
-    },
-  },
-}
-
-// Staggered list item animation
-export const listItem: Variants = {
-  hidden: {
-    opacity: 0,
-    y: 10,
-  },
-  show: {
-    opacity: 1,
-    y: 0,
-    transition: {
-      type: 'spring',
-      damping: 25,
-      stiffness: 200,
-    },
-  },
-}
-
-// Page transition variants
-export const pageVariants: Variants = {
-  initial: {
-    opacity: 0,
-  },
-  animate: {
-    opacity: 1,
-    transition: {
-      duration: 0.3,
-      ease: 'easeInOut',
-    },
-  },
-  exit: {
-    opacity: 0,
-    transition: {
-      duration: 0.2,
-      ease: 'easeInOut',
-    },
-  },
-}
-
-// Card hover animation
-export const cardHover = {
-  rest: {
-    scale: 1,
-    y: 0,
-    transition: {
-      duration: 0.2,
-      type: 'tween',
-      ease: 'easeOut',
-    },
-  },
-  hover: {
-    scale: 1.02,
-    y: -5,
-    transition: {
-      duration: 0.2,
-      type: 'tween',
-      ease: 'easeOut',
-    },
-  },
-}
-
-// Button hover animation
-export const buttonHover = {
-  rest: { scale: 1 },
-  hover: { scale: 1.05 },
-}
-
-// Gradient animation
-export const gradientAnimation: Variants = {
-  hidden: {
-    backgroundPosition: '0% 50%',
-  },
-  show: {
-    backgroundPosition: '100% 50%',
-    transition: {
-      repeat: Number.POSITIVE_INFINITY,
-      repeatType: 'reverse',
-      duration: 8,
-      ease: 'linear',
-    },
-  },
-}
